feat: refactor func
